% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcModMoransI.R
\name{calcModMoransI}
\alias{calcModMoransI}
\title{calcModMoransI}
\usage{
calcModMoransI(
  spe,
  l_prop = 0.1,
  weights_min = 0.01,
  x_coord = "pxl_row_fullres",
  y_coord = "pxl_col_fullres",
  verbose = FALSE
)
}
\arguments{
\item{spe}{Input object (SpatialExperiment). Assumed to contain an assay
named "logcounts" containing log-transformed normalized counts in sparse
matrix format, and "spatialCoords" slot containing spatial coordinates.}

\item{l_prop}{Value to set characteristic length parameter in squared
exponential kernel used to calculate weights matrix. The characteristic
length parameter is set to "l_prop" times the maximum range of the x or y
coordinates. Default = 0.1.}

\item{weights_min}{Minimum weights threshold. Weights (in the spatial
covariance matrix) that are below "weights_min" times the maximum weights
value are assumed to be zero.}

\item{x_coord}{Name of column in spatialCoords slot containing x-coordinates.
Default = "pxl_row_fullres".}

\item{y_coord}{Name of column in spatialCoords slot containing x-coordinates.
Default = "pxl_col_fullres".}

\item{verbose}{Whether to print messages. Default = FALSE.}
}
\value{
Returns a list containing output values (one value per gene).
}
\description{
Calculate modified Moran's I statistic to rank spatially variable genes
(SVGs).
}
\details{
Fast implementation of modified Moran's I statistic for ranking spatially
variable genes (SVGs).

We modify the definition of Moran's I statistic to make two
sparsity-preserving assumptions, due to the sparse nature of spatial
transcriptomics data and to speed up runtime.

- We assume that most genes are not detected in most spots (spatial
coordinates), and perform calculations using only the values from the
non-zero spots. For example, mean expression of each gene (which is used
inside the Moran's I formula) is calculated as the mean of the non-zero
spots.

- We assume that weights (in the spatial covariance matrix) below some
threshold (e.g. below 1% of the maximum weights value) are equal to zero.
}
\examples{
# to do

}
