% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClustersDimRed.R
\name{plotClustersDimRed}
\alias{plotClustersDimRed}
\title{plotClustersDimRed}
\usage{
plotClustersDimRed(
  spe,
  type = "UMAP",
  x_axis = "UMAP1",
  y_axis = "UMAP2",
  cluster_id = "cluster_id",
  palette = "libd_layer_colors"
)
}
\arguments{
\item{spe}{Input object (SpatialExperiment or SingleCellExperiment).}

\item{type}{Type of dimension reduction to use. Options are "PCA" or "UMAP".
Default = "UMAP".}

\item{x_axis}{Name of column in reducedDim to use for x-axis. Default =
"UMAP1".}

\item{y_axis}{Name of column in reducedDim to use for y-axis. Default =
"UMAP2".}

\item{cluster_id}{Name of column containing cluster IDs. Default =
"cluster_id".}

\item{palette}{Color palette. Available options are "libd_layer_colors" and
"Okabe-Ito". To use a custom palette, provide a vector of hex codes for the
colors.}
}
\value{
Returns ggplot object containing plot. Returning the plot as an
  object allows users to add additional ggplot elements (e.g. new title or
  different formatting).
}
\description{
Plots for clustering
}
\details{
Functions to generate plots for clustering results.

This function generates a plot showing cluster labels on the spots in the
dimension-reduced space (PCA or UMAP).
}
\examples{
TO DO

}
