% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClustersSpots.R
\name{plotClustersSpots}
\alias{plotClustersSpots}
\title{plotClustersSpots}
\usage{
plotClustersSpots(
  spe,
  x_coord = x_coord,
  y_coord = y_coord,
  cluster_id = cluster_id,
  palette = "libd_layer_colors"
)
}
\arguments{
\item{spe}{Input object (SpatialExperiment or SingleCellExperiment).}

\item{x_coord}{Name of column in colData containing x coordinates. Default =
x_coord. Note that the column name needs to be provided as a variable name,
not a character string (due to the way ggplot2 works).}

\item{y_coord}{Name of column in colData containing y coordinates. Default =
y_coord. Note that the column name needs to be provided as a variable name,
not a character string (due to the way ggplot2 works).}

\item{cluster_id}{Name of column containing cluster IDs. Default =
cluster_id. Note that the column name needs to be provided as a variable
name, not a character string (due to the way ggplot2 works).}

\item{palette}{Color palette. Available options are "libd_layer_colors" and
"Okabe-Ito". To use a custom palette, provide a vector of hex codes for the
colors.}
}
\value{
Returns ggplot object containing plot. Returning the plot as an
  object allows users to add additional ggplot elements (e.g. new title or
  different formatting).
}
\description{
Plots for clustering
}
\details{
Functions to generate plots for clustering results.

This function generates a plot showing cluster labels on the spots in the
physical x-y coordinates of the tissue slide.
}
\examples{
TO DO

}
