% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQCbar.R
\name{plotQCbar}
\alias{plotQCbar}
\title{plotQCbar}
\usage{
plotQCbar(spe, metric_x = "cell_count", highlight_zeros = TRUE)
}
\arguments{
\item{spe}{Input object (SpatialExperiment).}

\item{metric_x}{Name of column in colData containing QC metric to plot on
x-axis (e.g. "cell_count" for number of cells per spot). Default =
"cell_count".}

\item{highlight_zeros}{Whether to highlight bar for x = 0 (e.g. zero cells
per spot, which is a special value).}
}
\value{
Returns a ggplot object. Additional plot elements can be added as
  ggplot elements (e.g. title, formatting).
}
\description{
Quality control (QC) plots for spatial transcriptomics datasets.
}
\details{
Functions to generate quality control (QC) plots for spatial transcriptomics
datasets.

This function generates a barplot for one quality control (QC) metric, e.g.
number of cells per spot. For number of cells per spot, the barplot
highlights spots with zero cells, which are special values.
}
\examples{
# to do

}
