% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQCscatter.R
\name{plotQCscatter}
\alias{plotQCscatter}
\title{plotQCscatter}
\usage{
plotQCscatter(
  spe,
  metric_x = "cell_count",
  metric_y = "detected",
  threshold_x = NULL,
  threshold_y = NULL,
  trend = TRUE,
  marginal = TRUE
)
}
\arguments{
\item{spe}{Input object (SpatialExperiment or SingleCellExperiment).}

\item{metric_x}{Name of column in colData containing QC metric to plot on
x-axis (e.g. "cell_count" for number of cells per spot). Default =
"cell_count".}

\item{metric_y}{Name of column in colData containing QC metric to plot on
y-axis (e.g. "sum" for total library size, "detected" for number of
expressed features). Default = "detected".}

\item{threshold_x}{If provided, a vertical line will be drawn at this
x-value. Default = NULL.}

\item{threshold_y}{If provided, a horizontal line will be drawn at this
y-value. Default = NULL.}

\item{trend}{Whether to include a smoothed trend (loess). Default = TRUE.}

\item{marginal}{Whether to include marginal histograms. Default = TRUE.}
}
\value{
Returns a ggplot object. Additional plot elements can be added as
  ggplot elements (e.g. title, formatting).
}
\description{
Plots for quality control (QC)
}
\details{
Functions to generate plots for quality control (QC) purposes.

This function generates a scatterplot comparing two QC metrics, e.g. number
of detected features vs. number of cells per spot.
}
\examples{
# to do

}
