% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQCspots.R
\name{plotQCspots}
\alias{plotQCspots}
\title{plotQCspots}
\usage{
plotQCspots(spe, x_coord = x_coord, y_coord = y_coord, discard = discard)
}
\arguments{
\item{spe}{Input object (SpatialExperiment or SingleCellExperiment).}

\item{x_coord}{Name of column in colData containing x coordinates. Default =
x_coord. Note that the column name needs to be provided as a variable name,
not a character string (due to the way ggplot2 works).}

\item{y_coord}{Name of column in colData containing y coordinates. Default =
y_coord. Note that the column name needs to be provided as a variable name,
not a character string (due to the way ggplot2 works).}

\item{discard}{Name of column containing logical entries identifying spots to
be discarded. Default = discard. Note that the column name needs to be
provided as a variable name, not a character string (due to the way ggplot2
works).}
}
\value{
Returns ggplot object containing plot. Returning the plot as an
  object allows users to add additional ggplot elements (e.g. new title or
  different formatting).
}
\description{
Plots for quality control (QC)
}
\details{
Functions to generate plots for quality control (QC) purposes.

This function generates a plot showing spots to be discarded on the physical
x-y coordinates of the tissue slide.
}
\examples{
TO DO

}
