% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpotQC.R
\name{plotSpotQC}
\alias{plotSpotQC}
\title{plotSpotQC}
\usage{
plotSpotQC(
  sce,
  sample_name = "151673",
  plot_type = c("UMI", "gene"),
  save_plot = FALSE,
  filename = NULL,
  width = 5,
  height = 5
)
}
\arguments{
\item{sce}{Input object (SingleCellExperiment). Currently assumed to have the
structure from our DLPFC collaboration, i.e. specific rowData and colData
slots. Will generalize it later.}

\item{sample_name}{Which sample to subset from the SingleCellExperiment.
Assumes structure from our DLPFC collaboration, e.g. "151673" for sample
151673.}

\item{plot_type}{Plot type: either "UMI" or "gene", to plot either total UMI
counts per spot or total gene counts per spot vs. number of cells per spot.}

\item{save_plot}{Whether to save plot. If FALSE, plot will be displayed to
screen instead.}

\item{filename}{Filename and path to save plot.}

\item{width}{Plot width (inches).}

\item{height}{Plot height (inches).}
}
\value{
Saves a plot.
}
\description{
Plots for spot-level quality control (QC)
}
\details{
Plotting functions for spot-level quality control (QC)
}
\examples{
# load("../../../K99_grant_application/data/Human_DLPFC_Visium_processedData_sce_scran.Rdata")
# plotSpotQC(sce, sample_name = "151673")

}
