% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpots.R
\name{plotSpots}
\alias{plotSpots}
\title{plotSpots}
\usage{
plotSpots(
  spe,
  x_coord = "pxl_row_fullres",
  y_coord = "pxl_col_fullres",
  cluster_id = "cluster_id",
  palette = "libd_layer_colors"
)
}
\arguments{
\item{spe}{Input object (SpatialExperiment).}

\item{x_coord}{Name of column in spatialCoords slot containing x-coordinates.
Default = "pxl_row_fullres".}

\item{y_coord}{Name of column in spatialCoords slot containing x-coordinates.
Default = "pxl_col_fullres".}

\item{cluster_id}{Name of column in colData containing cluster IDs. To plot
without cluster labels, set "cluster_id = NULL". Default = "cluster_id".}

\item{palette}{Color palette for cluster labels. Available options are
"libd_layer_colors" and "Okabe-Ito". To use a custom palette, provide a
vector of hex codes for the colors.}
}
\value{
Returns a ggplot object. Additional plot elements can be added as
  ggplot elements (e.g. title, formatting).
}
\description{
Plotting functions for spatial transcriptomics datasets.
}
\details{
Functions to generate plots for spatial transcriptomics datasets.

This function generates a plot showing spatial coordinates (spots) in
physical x-y coordinates of the tissue slide. Cluster labels or ground truth
labels can be shown with colors.
}
\examples{
# to do

}
