% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessSVGs.R
\name{preprocessSVGs}
\alias{preprocessSVGs}
\title{preprocessSVGs}
\usage{
preprocessSVGs(
  spe,
  in_tissue = TRUE,
  filter_genes = 10,
  filter_mito = TRUE,
  seed = 123
)
}
\arguments{
\item{spe}{\code{SpatialExperiment} Input data, assumed to be a
\code{SpatialExperiment} object.}

\item{in_tissue}{\code{logical} Whether to keep only spots over tissue,
identified by column \code{in_tissue} in \code{spatialData}. Default =
TRUE.}

\item{filter_genes}{\code{integer} Filtering on UMI counts per gene. Genes
with less than or equal to this number of total UMI counts across spots
will be filtered out. Default = 10.}

\item{filter_mito}{\code{logical} Whether to filter out mitochondrial genes.
Assumes \code{rowData} contains a column named \code{gene_name}. Default =
TRUE.}

\item{seed}{\code{integer} Random seed for steps requiring random seed.
Default = 123.}
}
\value{
Returns a \code{SpatialExperiment} object that is ready to be
  provided as the input to SVGs methods.
}
\description{
Preprocessing steps to run SVGs methods
}
\details{
Convenience function to run several preprocessing steps to prepare data for
input to SVGs methods. Mainly intended for development work, but may also be
useful for others. The steps are intended for 10x Genomics Visium data. Code
is sourced from our online book \code{OSTA}, and makes use of the
\code{SpatialExperiment} structure.
}
\examples{
# to do

}
