% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQCscatter.R
\name{plotQCscatter}
\alias{plotQCscatter}
\title{plotQCscatter}
\usage{
plotQCscatter(
  spe,
  cell_count = "cell_count",
  metric = "detected",
  threshold = NULL,
  trend = FALSE,
  marginal = FALSE
)
}
\arguments{
\item{spe}{Input object (SpatialExperiment or SingleCellExperiment).}

\item{cell_count}{Name of column in colData containing number of cells per
spot to be plotted on x axis. Default = "cell_count".}

\item{metric}{Name of column in colData containing QC metric to be plotted on
y axis. Common options are "sum" (usually refers to total library size) and
"detected" (usually refers to number of expressed features). Default =
"detected".}

\item{threshold}{If provided, a horizontal line will be drawn at this
threshold for the metric.}

\item{trend}{Whether to include smoothed trend (loess). Default = FALSE.}

\item{marginal}{Whether to include marginal histograms. Default = FALSE.}
}
\value{
Returns ggplot object containing plot. Returning the plot as an
  object allows users to add additional ggplot elements (e.g. new title or
  different formatting).
}
\description{
Plots for quality control (QC)
}
\details{
Functions to generate plots for quality control (QC) purposes.

This function generates a scatterplot showing a QC metric (e.g. library size
or number of expressed features) vs. number of cells per spot.
}
\examples{
TO DO

}
