% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQCspots.R
\name{plotQCspots}
\alias{plotQCspots}
\title{plotQCspots}
\usage{
plotQCspots(spe, x_coord = "x_coord", y_coord = "y_coord", discard = "discard")
}
\arguments{
\item{spe}{Input object (SingleCellExperiment).}

\item{x_coord}{Name of column in colData containing x-coordinates. Default =
"x_coord".}

\item{y_coord}{Name of column in colData containing y-coordinates. Default =
"y_coord".}

\item{discard}{Name of column in colData identifying spots to be discarded
(TRUE/FALSE values). Default = "discard".}
}
\value{
Returns a ggplot object. Additional plot elements can be added as
  ggplot elements (e.g. title, formatting).
}
\description{
Plots for quality control (QC)
}
\details{
Functions to generate plots for quality control (QC) purposes.

This function generates a plot identifying spots that do not meet QC metrics
(i.e. spots to be discarded before downstream analysis) on the physical x-y
coordinates of the tissue slide.
}
\examples{
# to do

}
