% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankSVGsBRISC.R
\name{rankSVGsBRISC}
\alias{rankSVGsBRISC}
\title{rankSVGsBRISC}
\usage{
rankSVGsBRISC(
  spe,
  x = NULL,
  n.neighbors = 15,
  filter_counts = 5,
  filter_mito = TRUE,
  gene_name = "gene_name",
  n_threads = 4,
  ...
)
}
\arguments{
\item{spe}{Input object (SpatialExperiment). Assumed to contain "logcounts"
assay and spatial coordinates retrievable as "spatialCoords".}

\item{x}{Matrix of covariates for "BRISC_estimation()". Number of rows must
equal number of spots. See "?BRISC_estimation" for details. Default = NULL
(intercept-only model).}

\item{n.neighbors}{Number of nearest neighbors for BRISC. Default = 15.}

\item{filter_counts}{Filter out zero-expressed and low-expressed genes by
keeping genes with at least this number of UMI counts in at least one spot.
Default = 5.}

\item{filter_mito}{Whether to filter out mitochondrial genes. (Mitochondrial
genes are very highly expressed and may rank highly as SVGs if included,
but are often not of primary biological interest.) Default = TRUE.}

\item{gene_name}{Column in "rowData" containing gene names if "filter_mito =
TRUE". Default = "gene_name".}

\item{n_threads}{Number of threads for parallelization. Default = 4.}

\item{...}{Additional arguments to pass to "BRISC_estimation()".}
}
\value{
Returns test statistics and rankings of detected SVGs as new columns
  in "rowData". Genes that were not tested due to filtering (zero expression,
  low expression, and/or mitochondrial genes) are given NA values.
}
\description{
Rank spatially variable genes (SVGs) using BRISC approach
}
\details{
Calculate ranking of spatially variable genes (SVGs) using the BRISC approach
(bootstrap for rapid inference on spatial covariances) developed by Saha and
Datta (2018).
}
\examples{
# to do

}
