% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankSVGsBRISC.R
\name{runSVGsBRISC}
\alias{runSVGsBRISC}
\title{runSVGsBRISC}
\usage{
runSVGsBRISC(spe, x = NULL, n_threads = 1, ...)
}
\arguments{
\item{spe}{\code{SpatialExperiment} Input object, assumed to be a
\code{SpatialExperiment} containing an assay named \code{logcounts} and
spatial coordinates accessible with \code{spatialCoords()}.}

\item{x}{\code{numeric matrix} Matrix of covariates, with number of rows
(spots) matching the number of columns (spots) in \code{spe}. Default =
NULL, which is an intercept-only model. See \code{BRISC} documentation for
more details.}

\item{n_threads}{\code{integer} Number of threads for parallelization.
Default = 1.}

\item{...}{Additional arguments to pass to the \code{BRISC} estimation
function \code{BRISC_estimation()}.}
}
\value{
Returns summary statistics and SVG ranks as new columns in
  \code{rowData} in \code{spe} object.
}
\description{
Run method to identify spatially variable genes (SVGs) using BRISC
}
\details{
Identify top SVGs using BRISC ("bootstrap for rapid inference on spatial
covariances") methodology developed by Saha and Datta (2018).

This function runs BRISC separately for each gene, using parallelization for
faster runtime using one core per BRISC run. The main outputs of interest are
the parameter estimates of the covariance model stored in 'Theta' in the
BRISC output. We use the summary value 'sigma.sq / tau.sq' calculated from
these estimates to rank genes as SVGs, which represents spatial variance
divided by non-spatial variance. Alternatively, 'sigma.sq / (sigma.sq +
tau.sq)' could also be used, representing the proportion of spatial variance
out of total variance.

This version does not include bootstrap inference on the parameter estimates,
since this is much slower.

Assumes the input object is a \code{SpatialExperiment} containing an assay
named \code{logcounts} and filtered to exclude low-expressed genes, e.g. as
prepared with \code{\link{preprocessSVGs}}.
}
\examples{
# to do

}
